Revert "Revert "Reminder model""
